Fixed major resolution independence bug, added partial collision detection and added support for multiple assets per entity

+++ 2DGameEngine/Position.cpp
@@ -9,7 +9,7 @@ Position::Position()
 }
 
 
-Position::Position(double _x, double _y, double _z)
+Position::Position(int _x, int _y, int _z)
 {
 	X = _x;
 	Y = _y;
