+++ 2DGameEngine/Position.h
@@ -1,6 +1,8 @@
 #ifndef POSITION_H
 #define POSITION_H
 
+#include "Logger.h"
+
 /// <summary>
 /// A position is a simple model that contains X, Y and Z-values.
 /// </summary>
@@ -8,12 +10,12 @@ class Position
 {
 public:
 	Position();
-	Position(double, double, double);
+	Position(int, int, int);
 	~Position();
 
-	double X;
-	double Y;
-	double Z;
+	int X;
+	int Y;
+	int Z;
 };
 
 #endif
